requirements.txt
aiohttp==3.7.4.post0
backoff==1.11.1
discord.py==1.7.3
python-dateutil==2.8.2
pytz==2021.1
redis==3.5.3
Logbook==1.5.3
nose
freezegun==1.1.0
